chore: remove old FormProvider code
