+++ packages/widget/src/components/AmountInput/AmountInput.tsx
@@ -4,8 +4,9 @@ import type { ChangeEvent, ReactNode } from 'react';
 import { useLayoutEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useToken } from '../../hooks';
-import type { FormTypeProps } from '../../providers';
-import { FormKeyHelper, useWidgetConfig } from '../../providers';
+import type { FormTypeProps } from '../../stores';
+import { FormKeyHelper } from '../../stores';
+import { useWidgetConfig } from '../../providers';
 import { DisabledUI } from '../../types';
 import { fitInputText, formatInputAmount } from '../../utils';
 import { Card, CardTitle } from '../Card';
