+++ packages/widget/src/components/AmountInput/FormPriceHelperText.tsx
@@ -2,10 +2,9 @@ import type { TokenAmount } from '@lifi/sdk';
 import { FormHelperText, Skeleton, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { useTokenAddressBalance } from '../../hooks';
-import type { FormTypeProps } from '../../providers';
-import { FormKeyHelper } from '../../providers';
+import type { FormTypeProps } from '../../stores';
 import { formatTokenAmount, formatTokenPrice } from '../../utils';
-import { useFieldValues } from '../../stores';
+import { useFieldValues, FormKeyHelper } from '../../stores';
 
 export const FormPriceHelperText: React.FC<FormTypeProps> = ({ formType }) => {
   const [chainId, tokenAddress] = useFieldValues(
