+++ packages/widget/src/components/ChainSelect/ChainSelect.tsx
@@ -3,9 +3,8 @@ import type { EVMChain } from '@lifi/sdk';
 import { Avatar, Box, Skeleton, Tooltip, Typography } from '@mui/material';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import type { FormTypeProps } from '../../providers';
-import { FormKeyHelper } from '../../providers';
-import { maxChainToOrder, useFieldValues } from '../../stores';
+import type { FormTypeProps } from '../../stores';
+import { maxChainToOrder, useFieldValues, FormKeyHelper } from '../../stores';
 import { navigationRoutes } from '../../utils';
 import { ChainCard, ChainContainer } from './ChainSelect.style';
 import { useChainSelect } from './useChainSelect';
