+++ packages/widget/src/components/SendToWallet/SendToWalletButton.tsx
@@ -6,7 +6,6 @@ import { useRequiredToAddress, useWidgetEvents } from '../../hooks';
 import { useWidgetConfig } from '../../providers';
 import {
   useFieldActions,
-  useFormStore,
   useSendToWalletStore,
   useSettings,
 } from '../../stores';
