+++ packages/widget/src/components/TokenList/TokenNotFound.tsx
@@ -2,9 +2,8 @@ import SearchOffIcon from '@mui/icons-material/SearchOff';
 import { Box, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { useAvailableChains } from '../../hooks';
-import type { FormTypeProps } from '../../providers';
-import { FormKeyHelper } from '../../providers';
-import { useFieldValues } from '../../stores';
+import type { FormTypeProps } from '../../stores';
+import { useFieldValues, FormKeyHelper } from '../../stores';
 
 export const TokenNotFound: React.FC<FormTypeProps> = ({ formType }) => {
   const { t } = useTranslation();
