+++ packages/widget/src/hooks/useRoutes.ts
@@ -7,8 +7,8 @@ import { isAddress as isEVMAddress, parseUnits } from 'viem';
 import { normalize } from 'viem/ens';
 import { useConfig } from 'wagmi';
 import { useChain, useDebouncedWatch, useGasRefuel, useToken } from '.';
-import { FormKey, useWidgetConfig } from '../providers';
-import { useFieldValues, useSettings } from '../stores';
+import { useWidgetConfig } from '../providers';
+import { useFieldValues, useSettings, FormKey } from '../stores';
 import { isSVMAddress } from '../utils';
 import { useAccount } from './useAccount';
 import { useSwapOnly } from './useSwapOnly';
