+++ packages/widget/src/pages/SelectChainPage/types.ts
@@ -1,4 +1,4 @@
-import type { FormTypeProps } from '../../providers';
+import type { FormTypeProps } from '../../stores';
 
 export interface SelectChainPageProps extends FormTypeProps {
   selectNativeToken?: boolean;
