+++ packages/widget/src/pages/SelectNativeTokenPage/SelectNativeTokenPage.tsx
@@ -5,7 +5,7 @@ import { ListItemText } from '../../components/ListItemText';
 import { TokenAvatar } from '../../components/TokenAvatar';
 import { useTokenSelect } from '../../components/TokenList';
 import { useChains, useNavigateBack } from '../../hooks';
-import type { FormTypeProps } from '../../providers';
+import type { FormTypeProps } from '../../stores';
 
 export const SelectNativeTokenPage: React.FC<FormTypeProps> = ({
   formType,
