--- packages/widget/src/providers/FormProvider/index.ts
@@ -1,2 +1,0 @@
-export * from './FormProvider';
-export * from './types';
