+++ packages/widget/src/providers/index.ts
@@ -1,4 +1,3 @@
-export * from './FormProvider';
 export * from './I18nProvider';
 export * from './ThemeProvider';
 export * from './WalletProvider';
